feat(label): Label support for vertical PayPal button (#696)
* feat(label): Label support for vertical PayPal button

Buttons will now display labels in the vertical lahyout which have `allowPrimaryVertical` set to
`true` in the button config. Right now this is limited to the PayPal button labels
(PAYPAL/CHECKOUT/PAY/BUYNOW)

* style(no-trailing-spaces): Removing extra space after `determineCanRenderLabel` function

* fix(flow-type): Changing `label` to be of BUTTON_LABEL type in `renderButton`

Changing the `label` argument to be of BUTTON_LABEL type in `renderButton` so it the type matches
when it is passed to `determineCanRenderLabel`. Otherwise flow type check fails.

* test(button): Tweaking button layout tests for label changes

The button layout tests will no longer expect failure for labels on vertical PayPal buttons

* Triggering travis build

* fix(label): Labels only show for multiple buttons in vertical layout

Make sure that the layout is vertical to show labels for contexts where there are multiple buttons.

* refactor(determine-label): Changing boolean logic for `determineCanRenderLabel` into a fn that retur