feat(credit): PayPal Credit button will use DE logo for DE locales (#697)
* feat(credit): PayPal Credit button will use DE logo for DE locales

Adding in the DE PPC logo ("Ratenzahlung Powered by PayPal"). The `logoLabel` will use this logo in
the template if passed the DE locale.  Perhaps it should be separated out from credit into its own
funding option?

* style(missing-semicolon): Fixing missing semicolon causing test failure

* fix(constants): Adding the SOFORT constant back which got removed in the merge

* test(screenshots): Adding new screenshot tests for German PPC buttons

* fix(ppc-de-logo): Adjusting the logo viewbox/scale to not get cut off

* refactor(button-config): Extracting locale-label logic into individual configuration objects

* fix(label-config): Adding German PPC label to the localized label config.

This way it will show up when the merchant specifies the "credit" logo under the german locale.

* refactor(funding-logos): Moving PayPal Credit logo into its own jsx template

* chore(ci): trigger travis

* refactor(button-config): Reverting the `getLocalizedLabels` logic.

Keeping the German label logic in the config function, since this is an edge case.

* fix(localization): Updating the credit label text for Germany