Hide the menu-button under 250px
At under 250px we restrict the menu dropdowns. Some clients were running
into issues with the button appearing when it should not and causing a
bad user experience.

This fix is a bandaid to hide the button as communication between
smart-paymant-buttons and checkout-components does need some work to be
able to accurately determine when to disable rendering of menu in smart-payment-buttons.
On the brightside we would likely want this either way for responsive
behavior

[DTPPSDK-367]